[UsdGeom] Fix signed-vs-unsigned error.
(Internal change: 1712666)

+++ pxr/usd/lib/usdGeom/pointInstancer.cpp
@@ -733,7 +733,8 @@ UsdGeomPointInstancer::ComputeExtentAtTime(
     // verify that all the protoIndices are in bounds.
     TF_FOR_ALL(iter, protoIndices) {
         const int protoIndex = *iter;
-        if (protoIndex < 0 || protoIndex >= protoPaths.size()) {
+        if (protoIndex < 0 || 
+            static_cast<size_t>(protoIndex) >= protoPaths.size()) {
             TF_WARN("%s -- invalid prototype index: %d. Should be in [0, %zu)",
                     GetPrim().GetPath().GetText(),
                     protoIndex,
